requirements.txt
aioredis==1.3.1
asgiref==3.3.1
astroid==2.5.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.2
drf-yasg==1.20.0
firebase-admin==4.5.3
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-secret-manager==2.3.0
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.1
gunicorn==20.0.4
hiredis==2.0.0
httplib2==0.19.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
libcst==0.3.17
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
supervisor==4.2.2
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.3.0
